+++ apps/user/src/app/home/components/AgeSection.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import SectionHeader from './ui/SectionHeader';
-import DynamicCard from './ui/DynamicCard';
+import SectionHeader from '../../../components/ui/SectionHeader';
+import DynamicCard from '../../../components/ui/DynamicCard';
 import { BrandContent } from '@/types/brand';
 
 const AgeSection = () => {
